frontend/requirements.txt
numpy == 1.23.5
python-dotenv == 1.0.0
python-telegram-bot == 20.3
requests == 2.28.1